+++ Robust.Shared/Physics/Systems/SharedPhysicsSystem.Contacts.cs
@@ -855,10 +855,10 @@ public abstract partial class SharedPhysicsSystem
 
 public record struct ContactEnumerator
 {
-    public static readonly ContactEnumerator Empty = new(null);
-
     private Dictionary<string, Fixture>.ValueCollection.Enumerator _fixtureEnumerator;
     private Dictionary<Fixture, Contact>.ValueCollection.Enumerator _contactEnumerator;
+
+    private bool _empty;
 
     /// <summary>
     /// Also include deleting contacts.
@@ -872,7 +872,7 @@ public record struct ContactEnumerator
 
         if (fixtures == null || fixtures.Fixtures.Count == 0)
         {
-            this = Empty;
+            _empty = true;
             return;
         }
 
@@ -883,6 +883,12 @@ public record struct ContactEnumerator
 
     public bool MoveNext([NotNullWhen(true)] out Contact? contact)
     {
+        if (_empty)
+        {
+            contact = null;
+            return false;
+        }
+
         if (!_contactEnumerator.MoveNext())
         {
             if (!_fixtureEnumerator.MoveNext())